Run application middlewares on routing errors
If a request is made to an URI which doesn’t resolve to a route,
Confetti will currently not execute a middleware. This makes it
impossible to add behavior for handling these soft errors gracefully or
log the request.

Therefore, this makes application middlewares run on routing errors by
storing the middlewares on the RouteCollection and setting them when
creating an error route.